Separated LogHelper.php into Log.php and LogFactory.php Added logger, removed _log_file from ApiOperationBase.php
