+++ lib/net/authorize/util/LogFactory.php
@@ -1,0 +1,14 @@
+<?php
+namespace net\authorize\util;
+
+class LogFactory
+{
+    private static $logger = NULL;
+    public static function getLog($classType){
+        if(NULL == self::$logger){
+            self::$logger = new Log();
+        }
+        return self::$logger;
+    }
+}
+?>
